requirements.txt
discord.py==2.5.2
audioop-lts; python_version>='3.13'
python-dotenv==1.1.0
asgiref==3.8.1
aiofiles==24.1.0
beautifulsoup4==4.13.4
duckduckgo-search==8.0.1
pymupdf==1.25.5
cryptography==44.0.2
g4f==0.5.3.2
